requirements.txt
openai>=1.50.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
fastapi>=0.108.0
uvicorn>=0.25.0
httpx>=0.26.0
tenacity>=8.2.3
gradio>=4.0.0
pillow>=10.0.0

